+++ src/uni_modules/lucky-ui/components/lk-form/lk-form-item.vue
@@ -17,7 +17,6 @@ import {
   type FormContext,
   type FormRule,
   type FormItemContext,
-  type ValidateError,
 } from './context';
 import { useLocale } from '../../composables/useLocale';
 import {